Updated to use localStorage
